+++ cpp/include/raft/spatial/knn/common.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2021-2022, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,35 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/**
+ * This file is deprecated and will be removed in a future release.
+ * Please use the ann_types.hpp version instead.
+ */
 
 #pragma once
 
-#include <raft/distance/distance_type.hpp>
-
-namespace raft::spatial::knn {
-
-/** The base for approximate KNN index structures. */
-struct index {
-};
-
-/** The base for KNN index parameters. */
-struct index_params {
-  /** Distance type. */
-  raft::distance::DistanceType metric = distance::DistanceType::L2Expanded;
-  /** The argument used by some distance metrics. */
-  float metric_arg = 2.0f;
-  /**
-   * Whether to add the dataset content to the index, i.e.:
-   *
-   *  - `true` means the index is filled with the dataset vectors and ready to search after calling
-   * `build`.
-   *  - `false` means `build` only trains the underlying model (e.g. quantizer or clustering), but
-   * the index is left empty; you'd need to call `extend` on the index afterwards to populate it.
-   */
-  bool add_data_on_build = true;
-};
-
-struct search_params {
-};
-
-};  // namespace raft::spatial::knn
+#include <raft/spatial/knn/common.hpp>
